+++ client/app/modules/navbar/contactmodal.controller.js
@@ -1,10 +1,11 @@
 export class ContactModalController {
   /*@ngInject*/
-  constructor($uibModalInstance, Auth, $log, ContactService) {
+  constructor($uibModalInstance, Auth, $log, ContactService, toast) {
     this.$uibModalInstance = $uibModalInstance;
     this.Auth = Auth;
     this.$log = $log;
     this.contactService = ContactService;
+    this.toast = toast;
     this.contact = {};
     Auth.getCurrentUser()
       .then(user => {
@@ -20,9 +21,15 @@ export class ContactModalController {
   messageSend(form) {
     // Now we have the form data in this.contact
     if(form.$valid) {
-      this.contactService.messageSend(this.contact);
+      this.contactService.messageSend(this.contact)
+        .then(() => this.toast({
+          duration  : 5000,
+          message   : 'Message sent.',
+          className : 'alert-success'
+        }));
+
       // Don't wait for the promise to be fulfilled because email sending is slow
-      this.$uibModalInstance.close();
+      this.$uibModalInstance.close(); 
     }
   }